Refactor PolygonProcessor to be moved to general_tools
PolygonProcessor has been moved to the general_tools directory. This commit updates all import paths of PolygonProcessor in various files to reflect this structural change in the project. This ensures the correct functioning of the code post-refactoring.
