+++ custom_tools/general_tools/partition_iterator.py
@@ -15,7 +15,7 @@ from custom_tools.decorators.timing_decorator import timing_decorator
 
 from input_data import input_n50
 from file_manager.n100.file_manager_buildings import Building_N100
-from custom_tools.polygon_processor import PolygonProcessor
+from custom_tools.general_tools.polygon_processor import PolygonProcessor
 from constants.n100_constants import N100_Symbology
 
 # THIS IS WORK IN PROGRESS NOT READY FOR USE YET
