+++ generalization/n100/building/data_preparation.py
@@ -11,7 +11,7 @@ from file_manager.n100.file_manager_buildings import Building_N100
 from env_setup import environment_setup
 from custom_tools.decorators.timing_decorator import timing_decorator
 from custom_tools.general_tools import custom_arcpy
-from custom_tools.polygon_processor import PolygonProcessor
+from custom_tools.general_tools.polygon_processor import PolygonProcessor
 from input_data import input_symbology
 from constants.n100_constants import N100_Symbology, N100_SQLResources, N100_Values
