+++ generalization/n100/building/point_resolve_building_conflicts.py
@@ -8,7 +8,7 @@ from input_data import input_n100
 from input_data.input_symbology import SymbologyN100
 from file_manager.n100.file_manager_buildings import Building_N100
 from constants.n100_constants import N100_Symbology, N100_Values
-from custom_tools.polygon_processor import PolygonProcessor
+from custom_tools.general_tools.polygon_processor import PolygonProcessor
 from env_setup import environment_setup
 
 # Importing timing decorator
